% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhtmlwidgets160.R
\name{testhtmlwidgets160-shiny}
\alias{testhtmlwidgets160-shiny}
\alias{testhtmlwidgets160Output}
\alias{renderTesthtmlwidgets160}
\title{Shiny bindings for testhtmlwidgets160}
\usage{
testhtmlwidgets160Output(outputId, width = "100\%", height = "400px")

renderTesthtmlwidgets160(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a testhtmlwidgets160}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using testhtmlwidgets160 within Shiny
applications and interactive Rmd documents.
}
