% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhtmlwidgets160.R
\name{testhtmlwidgets160}
\alias{testhtmlwidgets160}
\title{Title}
\usage{
testhtmlwidgets160(message, width = NULL, height = NULL, elementId = NULL)
}
\description{
Test sizing policy changes in htmlwidgets version 1.6.0
}
